Expose host capability namespaces
