+++ js/kubewarden/host_capabilities/crypto/crypto.test.ts
@@ -1,10 +1,10 @@
-import { HostCall } from '../';
+import { HostCall } from '../host_call';
 
 import { Crypto } from './crypto';
 import type { Certificate, CertificateVerificationResponse } from './types';
 
 // Mock the HostCall
-jest.mock('../', () => ({
+jest.mock('../host_call', () => ({
   HostCall: {
     hostCall: jest.fn(),
   },
