+++ js/kubewarden/host_capabilities/crypto/crypto.ts
@@ -1,4 +1,4 @@
-import { HostCall } from '../';
+import { HostCall } from '../host_call';
 
 import type { Certificate, CertificateVerificationResponse } from './types';
 import { CertificateUtils } from './types';
