+++ js/kubewarden/host_capabilities/net/network.ts
@@ -1,4 +1,4 @@
-import { HostCall } from '../index';
+import { HostCall } from '../host_call';
 
 export namespace Network {
   /**
