+++ js/kubewarden/host_capabilities/oci/manifest/manifest.ts
@@ -1,4 +1,4 @@
-import { HostCall } from '../..';
+import { HostCall } from '../../host_call';
 
 import { OciImageManifestResponse } from './types';
