+++ js/kubewarden/host_capabilities/oci/manifest_config/manifest_config.ts
@@ -1,4 +1,4 @@
-import { HostCall } from '../..';
+import { HostCall } from '../../host_call';
 
 import type { OciImageManifestAndConfigResponse } from './types';
