+++ js/kubewarden/host_capabilities/oci/manifest_digest/manifest_digest.ts
@@ -1,4 +1,4 @@
-import { HostCall } from '../..';
+import { HostCall } from '../../host_call';
 
 import type { OciManifestResponse } from './types';
