+++ js/kubewarden/host_capabilities/oci/verify_v2/verifier.ts
@@ -1,4 +1,4 @@
-import { HostCall } from '../..';
+import { HostCall } from '../../host_call';
 
 import type {
   SigstorePubKeyVerifyRequest,
